Proposed fix for pages issue
This fixes the problem where Hotaru would look in a theme templates
folder for a file that matched the pageName in the url. E.g.
example.com/footer would display the footer template in the body pf the
page, thinking it was a static page.

This fix creates a new "pages" folder with an example file (faq.php).
Visiting example.com/faq will display the new static page, but
example.com/footer will correctly result in page not found.

This fix uses the name "pages" in displayTemplate's "plugin" parameter
even though it isn't a plugin. That's okay though because any future
plugin called "pages" will still be compatible with this fix.